# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roommate Chore & Expense Roulette API",
    "version": "1.0.0",
    "description": "# 🎲 Roommate Chore & Expense Roulette API\n\nProduction-grade FastAPI backend featuring specialized Data Structures & Algorithms (DSA):\n- **Chore Rotation**: Circular Queue (Ring Buffer) with O(1) modulo turn shift and sarcastic alerts\n- **Expense Ledger**: Hash Tables with O(1) net balance computations and strict zero-sum conservation\n- **Debt Simplification**: Min-Cash-Flow Greedy Directed Graph Algorithm reducing O(N^2) debts to <= N - 1 vectors\n- **Activity Undo Stack**: LIFO Stack for idempotent database state reversal\n- **Identity & Keys**: RFC 9562 UUIDv7 monotonic primary keys and Ed25519 asymmetric cryptography",
    "contact": {
      "name": "Roulette Engineering Team",
      "email": "engineering@roommate-roulette.local",
      "url": "https://github.com/roommate-roulette"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.roommate-roulette.local",
      "description": "Production Gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user account",
        "description": "Generates an RFC 9562 UUIDv7 ID, hashes password with Bcrypt, and issues a signed PyJWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully and token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email already registered"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Bcrypt password login",
        "description": "Authenticates user email and password, issuing a signed PyJWT token with 7-day expiry.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/v1/auth/google": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Google OAuth2 credential login",
        "description": "Validates Google OAuth JWT ID token and links or creates user account seamlessly.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/link-google": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Link Google Auth as alternative login method to existing account",
        "description": "Attaches Google OAuth2 credentials to an existing user account. Validates that the Google identity is not already linked to another roommate and returns a fresh JWT with updated IAM claims.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkGoogleAuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google Auth linked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkGoogleAuthResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Google identity already linked to another account"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/v1/auth/unlink-google": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Unlink Google Auth from account (with lockout prevention)",
        "description": "Detaches Google login method from user account. Enforces lockout prevention: user must retain at least one alternative login method (password or Ed25519 key).",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlinkGoogleAuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google Auth unlinked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMethodsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Lockout prevention error or no Google account linked"
          }
        }
      }
    },
    "/api/v1/auth/methods/{user_id}": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Query configured authentication methods for roommate",
        "description": "Returns active authentication strategies: password, ed25519, and google auth.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "User UUID7",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authentication methods retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMethodsResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/api/v1/auth/ed25519-login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Ed25519 cryptographic signature login",
        "description": "Verifies a 64-byte Ed25519 Edwards-curve signature against the public key and server challenge nonce.",
        "responses": {
          "200": {
            "description": "Cryptographic signature verified and JWT issued"
          },
          "401": {
            "description": "Invalid cryptographic signature"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "PyJWT Bearer token"
      },
      "Ed25519Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Ed25519-Signature",
        "description": "64-byte Ed25519 signature over server challenge nonce"
      },
      "Ed25519PublicKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Ed25519-Public-Key",
        "description": "32-byte Ed25519 hex public key"
      }
    },
    "schemas": {
      "UserRegisterRequest": {
        "title": "UserRegisterRequest",
        "type": "object",
        "required": ["email", "password", "full_name"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "full_name": { "type": "string" },
          "ed25519_public_key": { "type": "string" }
        }
      },
      "UserLoginRequest": {
        "title": "UserLoginRequest",
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "TokenResponse": {
        "title": "TokenResponse",
        "type": "object",
        "required": ["access_token", "token_type", "user_id", "email", "full_name"],
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "default": "bearer" },
          "user_id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "full_name": { "type": "string" },
          "google_sub_id": { "type": "string" },
          "available_login_methods": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "GoogleAuthRequest": {
        "title": "GoogleAuthRequest",
        "type": "object",
        "required": ["credential_token"],
        "properties": {
          "credential_token": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "full_name": { "type": "string" },
          "target_user_id": { "type": "string", "format": "uuid" }
        }
      },
      "LinkGoogleAuthRequest": {
        "title": "LinkGoogleAuthRequest",
        "type": "object",
        "required": ["credential_token"],
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "credential_token": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "full_name": { "type": "string" }
        }
      },
      "LinkGoogleAuthResponse": {
        "title": "LinkGoogleAuthResponse",
        "type": "object",
        "required": ["status", "message", "user_id", "email", "full_name", "google_sub_id", "available_login_methods", "access_token", "token_type"],
        "properties": {
          "status": { "type": "string" },
          "message": { "type": "string" },
          "user_id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "full_name": { "type": "string" },
          "google_sub_id": { "type": "string" },
          "available_login_methods": {
            "type": "array",
            "items": { "type": "string" }
          },
          "access_token": { "type": "string" },
          "token_type": { "type": "string" }
        }
      },
      "UnlinkGoogleAuthRequest": {
        "title": "UnlinkGoogleAuthRequest",
        "type": "object",
        "properties": {
          "user_id": { "type": "string", "format": "uuid" }
        }
      },
      "AuthMethodsResponse": {
        "title": "AuthMethodsResponse",
        "type": "object",
        "required": ["user_id", "email", "full_name", "has_password", "has_ed25519", "has_google_auth", "available_login_methods"],
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "full_name": { "type": "string" },
          "has_password": { "type": "boolean" },
          "has_ed25519": { "type": "boolean" },
          "ed25519_public_key": { "type": "string" },
          "has_google_auth": { "type": "boolean" },
          "google_sub_id": { "type": "string" },
          "available_login_methods": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      }
    }
  }
}